Enable multi-architecture docker image builds
Previously, the Dockerfile downloaded 'docker-gen' and 'forego' binaries
during build time. This caused a problem as it hard-coded the amd64
architecture for the images.

This commit updates both 'Dockerfile' and 'Dockerfile.alpine' to build
the `forego` and `docker-gen` executables from scratch instead of
downloading binaries directly.

This is achieved using multi-stage builds [1]. Two seperate stages first
build the binaries, and are then copied over to the final stage.

The advantage of this change is two-fold: First, it enables building
this image on architectures other than amd64. Secondly it adds trust by
not adding external binaries to the docker image.

This modified version passes the test both a linux desktop (amd64) as
well as a raspberry pi (armv7) with some caveats:

- On armv7, a modified version of the `jwilder/docker-gen` image is
required. See a seperate PR at [2].

- The 'test_dhparam_is_generated_if_missing' test fails. This also
doesn't currently pass on master.

[1] https://docs.docker.com/develop/develop-images/multistage-build/ [2]
https://github.com/jwilder/docker-gen/pull/327